requirements.txt
asgiref==3.5.0
certifi==2021.10.8
click==8.1.3
cmake==3.22.4
colorama==0.4.6
distlib==0.3.4
filelock==3.6.0
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
nodejs==0.1.1
npm==0.1.1
numpy==1.22.3
optional-django==0.1.0
packaging==23.0
pdf2image==1.16.2
Pillow==9.1.0
pipenv==2022.4.21
platformdirs==2.5.2
pytesseract==0.3.10
six==1.16.0
sqlparse==0.4.2
torch==1.11.0
typing_extensions==4.2.0
tzdata==2022.1
virtualenv==20.14.1
virtualenv-clone==0.5.7
vue==0.0.1
Werkzeug==2.2.2

